chore: update German translations to state of v0.6.7 (#5848)
* chore: update translations with Fink 🐦

* chore: update translations with Fink 🐦

* fix: add empty itemSeven removed by Inlang

* fix: add empty itemSeven removed by Inlang

fix for turkish translation

* fix: empty translation for itemSeven

quickfix for the empty translation of itemSeven. This has been done the same way to resolve previous issues with InLang and empty translation strings.

* fix: empty translation for itemSeven in TR

quickfix for the empty translation of itemSeven. This has been done the same way to resolve previous issues with InLang and empty translation strings.

* fix: remove itemSeven completely

* chore: update translations with Fink 🐦

* chore: update translations with Fink 🐦

* fix: resolve conflicts after merging main

* chore: update translations with Fink 🐦

---------

Co-authored-by: Mathias Mogensen <mathiasrieckm@gmail.com>
Co-authored-by: Mathias Mogensen <mathias@appflowy.io>